requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
PyPDF2==3.0.1
google-generativeai==0.3.2
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1
